Put implicit transactions back.
Thinking about the way I've come up with this, it will be possible to
create transactions on top of transaction databases. What. So to prevent
endless loops, transaction databases need to have certain flags
enabled/disabled, and implicit_tx needs to be one of the ones turned
off.

If you really want to, enable them on your own.