ENGLISH A SHIT
Now it's time to taste on the first drops of how the birds must feel
like..... This is motherfucking Leaving Las Vegas !!!! Wiping the
memory N-O-W and see what they say when i'm the last one standing!!
When I first heard the H-word at 12 years old, it didn't sound like
a "positive"-sounding word as compared to "hopes" and "dreams". It
still doesn't sound like a positive word, but out of the sheer power
of it, it has become the most encouraging word in HER language,
surpassing everything agglutinative and beyond. And yet, these haters
STILL DON'T use the word amongst themselves in everyday speech. Isn't
that cute?

Do you feel the mother, are you afraid of flirting with her? Are you
afraid she might have a boyfriend? AFRAID she'll call the police on
violation or offense?

THAT IS NOT VERY P-O-S-I-T-I-V-E OF YOU??