fixes the builds for liboleg and the frontend
- ol_database: fixes length for config fields

Newer versions of gcc are very strict on the destination buffer length
in calls to strncpy. This commit ensures that the struct fields on
ol_database are long enough to account for the null terminator. This
commit also fixes the test functions to derive the analysis buffer from
the actual length of the db struct fields, ensuring we don't get
compiler warnings about strncpy.

- frontend: cleans up go package and adds a go.mod
- frontend: moves go main to ./cmd/olegdb and goleg package to
  ./pkg/goleg
- frontend: bumps go compiler version to 1.14